suppression des warning de rm en vérifiant l'existence des files

+++ main.sh
@@ -62,14 +62,14 @@ fi
 # crée l'arborescence de dossiers si elle n'existe pas
 mkdir -p $DB $RES/{blast,iadhore,python}
 
-
 # crée les bases de données de blast de PP et MD si leur dossier n'existe pas, ou est vide, ou le flag blast est levé
 
 # base de données MD
 if [[ blast_flag -eq 1 || ! -d "$DB/MD-db/" || ! "$(ls -A $DB/MD-db/)" ]]
 then
 	printf "\n### MAKEBLASTDB MD ###"
-	rm -r "$DB/MD-db/*"
+
+	if [[ -d "$DB/MD-db" && "$(ls -A $DB/MD-db)" ]] ; then rm -r $DB/MD-db/* ; fi # supprime l'ancienne db si elle existe
 	makeblastdb -in "$DATA/Malus-domestica-proteome.fasta" -dbtype prot -out "$DB/MD-db/MD-db"
 	printf "makeblastdb MD: done\n"
 fi
@@ -77,7 +77,8 @@ fi
 if [[ blast_flag -eq 1 || ! -d "$DB/PP-db/" || ! "$(ls -A $DB/PP-db/)" ]]
 then
 	printf "\n### MAKEBLASTDB PP ###"
-	rm -r "$DB/PP-db/*"
+
+	if [[ -d "$DB/PP-db" && "$(ls -A $DB/PP-db)" ]] ; then rm -r $DB/PP-db/* ; fi # supprime l'ancienne db si elle existe
 	makeblastdb -in "$DATA/Prunus-persica-proteome.fasta" -dbtype prot -out "$DB/PP-db/PP-db"
 	printf "makeblastdb PP: done\n"
 fi
@@ -87,7 +88,7 @@ fi
 if [[ blast_flag -eq 1 || ! "$(ls -A $RES/blast)" ]]
 then
 	printf "\n### BLAST ###\n"
-	rm -r "$RES/blast/*"
+	if [[ "$(ls -A $RES/blast)" ]] ; then rm -r $RES/blast/* ; fi
 	printf "DATA=$DATA\nDB=$DB\nOUT=$RES/blast\n"
 
 	# soumet le job blast
@@ -105,18 +106,22 @@ fi
 
 
 # remplit le fichier de config iadhore avec les chromosomes présents dans data
+# liste des fichiers de data/MD_lst de la forme :
+# filename1 this/is/the/path/filename1.extension
+# filename2 this/is/the/path/filename2.extension
+# filename3 this/is/the/path/filename3.extension
 make_iadhore_config() {
 	file="scripts/iadhore/iadhore.ini"
-	cat "scripts/iadhore/iadhore1_model.ini" > $file
+	cat "scripts/iadhore/iadhore1_model.ini" > $file # copie le modèle vide dans le fichier de config
 	
-	printf "\ngenome = Malus_domestica\n" >> $file
+	printf "\ngenome = Malus_domestica\n" >> $file 
 	ls -1 data_test/MD_lst/* | sed 's/.*/& &/' | sed -e 's/data_test\/MD_lst\///' -e 's/.lst / /' >> $file
-	
+	# lister deux fois les sorties séparées par un espace, puis supprimer le chemin et l'extension de la première sortie
+
 	printf "\ngenome = Prunus_persica\n" >> $file
 	ls -1 data_test/PP_lst/* | sed 's/.*/& &/' | sed -e 's/data_test\/PP_lst\///' -e 's/.lst / /' >> $file
 
-	cat $file
-
+	#cat $file # affiche le fichier de config
 }
 
 
@@ -125,13 +130,10 @@ if [[ iadhore_flag -eq 1 || ! "$(ls -A $RES/iadhore)" ]]
 then
 	printf "\n### IADHORE ###\n"
 
-	rm -r $RES/iadhore/* # supprime le contenu du dossier iadhore/
+	if [[ "$(ls -A $RES/iadhore)" ]] ; then rm -r $RES/iadhore/* ; fi # supprime les précédants resultats iadhore
 
 	# supprime le dossier tmp s'il existe puis en crée un vide
-	if [[ -d $TMP ]]
-	then
-		rm -r $TMP
-	fi
+	if [[ -d $TMP ]] ; then rm -r $TMP ; fi
 	mkdir -p $TMP/iadhore/
 
 	# déplace les fichiers d'entrée de DATA dans TMP/data pour que iadhore les choppe là-bas sans avoir à modifier son input path
@@ -146,9 +148,9 @@ then
 
 	make_iadhore_config # crée le fichier de config iadhore
 	
-	printf "\n=== AVANT SUBMIT : ===\n"
-	ls -R $TMP
-	printf "\n===\n"
+	#printf "\n=== AVANT SUBMIT : ===\n"
+	#ls -R $TMP
+	#printf "\n===\n"
 
 	# soumet le job iadhore
 	sbatch --wait $SUBMIT/iadhore_job.sh
@@ -163,11 +165,7 @@ then
 		sed 's:\t\t*:\t:g' $RES/iadhore/multiplicon_pairs.txt > $RES/iadhore/multiplicon_pairs_modified.txt
 	fi
 
-	# supprime le dossier temporaire
-	if [[ -d $TMP ]]
-	then
-		rm -r $TMP
-	fi
+	if [[ -d $TMP ]] ; then rm -r $TMP ; fi # supprime le dossier temporaire
 
 	printf "iadhore: done\n"
 fi
